+++ src/commands/del.js
@@ -6,18 +6,18 @@ const path = require("path");
 
 const _errorInterpret = require("../functions/errorInt");
 
-const rmfile = (fileNameParam) => {
-  if (typeof fileNameParam === "undefined") {
-    _errorInterpret("0x0035");
+const del = (file) => {
+  if (typeof file === "undefined") {
+    _errorInterpret("0x0029");
     return;
   }
 
-  const isAbsolutePath = path.isAbsolute(fileNameParam);
+  const isAbsolutePath = path.isAbsolute(file);
   let fileName = "";
   if (!isAbsolutePath) {
-    fileName = process.cwd() + "\\" + fileNameParam;
+    fileName = process.cwd() + "\\" + file;
   } else {
-    fileName = fileNameParam;
+    fileName = file;
   }
 
   if (!fs.existsSync(fileName)) {
@@ -27,31 +27,30 @@ const rmfile = (fileNameParam) => {
 
   const confirmText = readline
     .question(
-      `Are you sure you want to delete the file: ${chalk.bold(fileName)}? [${chalk.green(
+      `Are you sure you want to delete ${chalk.bold(fileName)}? [${chalk.green(
         "y"
       )}/${chalk.red.bold("N")}] `
     )
     .toLowerCase();
   if (confirmText.includes("n") || !confirmText.includes("y")) {
-    _errorInterpret("0x0037");
+    _errorInterpret("0x0031");
     return;
   }
 
-  console.log(`\nDeleting file: ${chalk.bold.blueBright(fileName)}...`);
+  console.log(`\nDeleting ${chalk.bold.blueBright(fileName)}...`);
 
   try {
-    fs.unlinkSync(fileName);
-
+    fs.rmSync(fileName, { recursive: true, force: true });
     console.log(chalk.green("The operation completed successfully.\n"));
   } catch (err) {
     if (err.code === "EBUSY") {
-      _errorInterpret("0x0038", { variable: fileName });
+      _errorInterpret("0x0032", { variable: fileName });
     } else if (err.code === "EPERM") {
-      _errorInterpret("0x0039", { variable: fileName });
+      _errorInterpret("0x0033", { variable: fileName });
     } else {
-      _errorInterpret("0x0040", { variable: fileName, wordCode: err.code });
+      _errorInterpret("0x0044", { variable: fileName, wordCode: err.code });
     }
   }
 };
 
-module.exports = rmfile;
+module.exports = del;